Add Toast message on delete article and folder #20

+++ apps/mdreader/src/pages/profile/useFolderAndArticleActions.ts
@@ -1,8 +1,10 @@
-import { useCallback, useMemo } from "react";
-import { File, Folder } from "lucide-react";
+import { useCallback, useMemo, useState } from 'react';
+import { File, Folder } from 'lucide-react';
 
-import { articleService } from "../../services/article";
-import { folderService } from "../../services/folder";
+import { useToast } from '@mdreader/ui/components/ui/use-toast';
+
+import { articleService } from '../../services/article';
+import { folderService } from '../../services/folder';
 
 type Props = {
   articles: any[];
@@ -11,31 +13,60 @@ type Props = {
   mutateFolders: any;
 };
 
+type ConfirmDialogState = {
+  open: boolean;
+  onConfirm: () => void;
+};
+
+const getDescriptionMessage = (
+  text: string
+) => `This action cannot be undone. This will permanently delete your
+"${text}".`;
+
 const useFolderAndArticleActions = ({
   articles,
   folders,
   mutateArticles,
   mutateFolders,
 }: Props) => {
+  const { toast } = useToast();
+
+  const [confirmDialogProps, setConfirmDialogProps] =
+    useState<ConfirmDialogState>({
+      open: false,
+      onConfirm: () => undefined,
+    });
+
   const onDelete = useCallback((record: any, mutate: any) => {
     mutate((prevRecords: any[]) =>
       prevRecords
         ? prevRecords?.filter((prevRecord) => prevRecord.id !== record.id)
         : []
     );
+
+    toast({ title: 'Success' });
   }, []);
 
-  const folderAndArticleItems = useMemo(() => {
+  const items = useMemo(() => {
     const _articles = articles.map((article) => ({
       ...article,
       actions: [
-        { name: "Edit", onClick: () => alert("Test") },
         {
-          name: "Remove",
+          name: 'Edit',
+          onClick: () => alert('Test'),
+        },
+        {
+          name: 'Remove',
           onClick: async () =>
-            articleService
-              .remove(article.id)
-              .then(() => onDelete(article, mutateArticles)),
+            setConfirmDialogProps((prevState) => ({
+              ...prevState,
+              description: getDescriptionMessage(article.name),
+              onConfirm: () =>
+                articleService
+                  .remove(article.id)
+                  .then(() => onDelete(article, mutateArticles)),
+              open: true,
+            })),
         },
       ],
       href: `preview/${article.id}`,
@@ -45,13 +76,19 @@ const useFolderAndArticleActions = ({
     const _folders = folders.map((folder) => ({
       ...folder,
       actions: [
-        { name: "Edit", onClick: () => alert("Test") },
+        { name: 'Edit', onClick: () => alert('Test') },
         {
-          name: "Remove",
+          name: 'Remove',
           onClick: async () =>
-            folderService
-              .remove(folder.id)
-              .then(() => onDelete(folder, mutateFolders)),
+            setConfirmDialogProps((prevState) => ({
+              ...prevState,
+              description: getDescriptionMessage(folder.name),
+              open: true,
+              onConfirm: () =>
+                folderService
+                  .remove(folder.id)
+                  .then(() => onDelete(folder, mutateFolders)),
+            })),
         },
       ],
       href: `?folderId=${folder.id}`,
@@ -59,9 +96,23 @@ const useFolderAndArticleActions = ({
     }));
 
     return [..._folders, ..._articles];
-  }, [articles, folders, onDelete, mutateArticles, mutateFolders]);
+  }, [
+    articles,
+    folders,
+    mutateArticles,
+    mutateFolders,
+    onDelete,
+    setConfirmDialogProps,
+  ]);
 
-  return folderAndArticleItems;
+  return {
+    confirmDialogProps: {
+      ...confirmDialogProps,
+      onCancel: () =>
+        setConfirmDialogProps((prevProps) => ({ ...prevProps, open: false })),
+    },
+    items,
+  };
 };
 
 export default useFolderAndArticleActions;
